recover by privkey (#69)
